atecc608a: Clarify that key IDs are not key slots
In showing this example to others not familiar with the PSA Crypto API,
this obstacle to learning became apparent: that it was confusing to
re-use the same variables, constants, and values to two different
concepts. Use separate values for key IDs and key slots, and add some
comments, to help show that the two concepts are different.