.travis.yml: Add NRF52_DK to test
By default, Mbed OS PSA is only enabled on K64F and K66F. This
example overrides the configuration to globally enable Mbed OS PSA
as needed.

We build NRF52_DK in Travis to verify this override. This target was
chosen because the pins for atecc608a are already provided in
`mbed_app.json`.